requirements.txt
groq
python-dotenv

# optional (voice)
gtts
pygame
SpeechRecognition
pyaudio